+++ analysis/tools/ValidationTool/src/main/kotlin/de/maibornwolff/codecharta/tools/validation/ValidationTool.kt
@@ -1,5 +1,7 @@
 package de.maibornwolff.codecharta.tools.validation
 
+import de.maibornwolff.codecharta.tools.interactiveparser.InteractiveParser
+import de.maibornwolff.codecharta.tools.interactiveparser.ParserDialogInterface
 import picocli.CommandLine
 import java.io.File
 import java.io.FileInputStream
@@ -10,9 +12,7 @@ import java.util.concurrent.Callable
     description = ["validates cc.json files"],
     footer = ["Copyright(c) 2020, MaibornWolff GmbH"]
 )
-class ValidationTool : Callable<Void?> {
-
-    private val schemaPath = "cc.json"
+class ValidationTool : Callable<Void?>, InteractiveParser {
 
     @CommandLine.Option(names = ["-h", "--help"], usageHelp = true, description = ["displays this help and exits"])
     var help: Boolean = false
@@ -21,17 +21,14 @@ class ValidationTool : Callable<Void?> {
     var file: String = ""
 
     override fun call(): Void? {
-        EveritValidator(schemaPath).validate(FileInputStream(File(file).absoluteFile))
+        EveritValidator(SCHEMA_PATH).validate(FileInputStream(File(file).absoluteFile))
 
         return null
     }
 
     companion object {
-        var SCHEMA_PATH = "cc.json"
-
-        @JvmStatic
-        fun main(args: Array<String>) {
-            CommandLine.call(ValidationTool(), System.out, *args)
-        }
+        const val SCHEMA_PATH = "cc.json"
     }
+
+    override fun getDialog(): ParserDialogInterface = ParserDialog
 }