+++ visualization/app/codeCharta/state/store/appSettings/amountOfEdgePreviews/amountOfEdgePreviews.selector.ts
@@ -1,0 +1,4 @@
+import { createSelector } from "../../../angular-redux/createSelector"
+import { appSettingsSelector } from "../appSettings.selector"
+
+export const amountOfEdgePreviewsSelector = createSelector([appSettingsSelector], appSettings => appSettings.amountOfEdgePreviews)