+++ visualization/app/codeCharta/state/store/fileSettings/edges/edges.selector.ts
@@ -1,0 +1,4 @@
+import { createSelector } from "../../../angular-redux/createSelector"
+import { fileSettingsSelector } from "../fileSettings.selector"
+
+export const edgesSelector = createSelector([fileSettingsSelector], fileSettings => fileSettings.edges)