+++ visualization/app/codeCharta/ui/toolBar/globalConfigurationButton/globalConfigurationDialog/displayQualitySelection/displayQualitySelection.module.ts
@@ -1,0 +1,11 @@
+import { CommonModule } from "@angular/common"
+import { NgModule } from "@angular/core"
+import { MaterialModule } from "../../../../../../material/material.module"
+import { DisplayQualitySelectionComponent } from "./displayQualitySelection.component"
+
+@NgModule({
+	imports: [CommonModule, MaterialModule],
+	declarations: [DisplayQualitySelectionComponent],
+	exports: [DisplayQualitySelectionComponent]
+})
+export class DisplayQualitySelectionModule {}